pyproject.toml
[project]
name = "campus-orbit-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.17.2",
    "asgi-lifespan>=2.1.0",
    "asyncio>=4.0.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.124.2",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.38.0",
]
[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
